Rename task, redo testnet deployments
