--- pkg/deployments/tasks/20210701-liquidity-bootstrapping-pool/input.ts
@@ -1,11 +1,0 @@
-import Task from '../../src/task';
-
-export type LiquidityBootstrappingPoolDeployment = {
-  vault: string;
-};
-
-const vault = new Task('20210418-vault');
-
-export default {
-  vault,
-};
