--- pkg/deployments/tasks/20210701-liquidity-bootstrapping-pool/output/kovan.json
@@ -1,4 +1,0 @@
-{
-  "factory": "0xdAE7e32ADc5d490a43cCba1f0c736033F2b4eFca",
-  "timestamp": 1625180322205
-}
