--- pkg/deployments/tasks/20210701-liquidity-bootstrapping-pool/output/rinkeby.json
@@ -1,4 +1,0 @@
-{
-  "factory": "0x72D07D7DcA67b8A406aD1Ec34ce969c90bFEE768",
-  "timestamp": 1625180465505
-}
