--- pkg/deployments/tasks/20210701-liquidity-bootstrapping-pool/readme.md
@@ -1,11 +1,0 @@
-# 2021-07-01 - Liquidity Bootstrapping Pool
-
-## 🚧 UNDER CONSTRUCTION - NOT PRODUCTION READY 🚧
-
-First testnet deployment of the factory for Liquidity Bootstrapping Pool, meant to be used for initial UI development.
-
-## Useful Files
-
-- [`LiquidityBootstrappingPool` ABI](./abi/LiquidityBootstrappingPool.json)
-- [Rinkeby testnet address](./output/rinkeby.json)
-- [Kovan testnet address](./output/kovan.json)
