+++ pkg/deployments/tasks/20210721-liquidity-bootstrapping-pool/output/kovan.json
@@ -1,0 +1,4 @@
+{
+  "factory": "0x0F3e0c4218b7b0108a3643cFe9D3ec0d4F57c54e",
+  "timestamp": 1626887237645
+}
