+++ pkg/deployments/tasks/20210721-liquidity-bootstrapping-pool/output/polygon.json
@@ -1,0 +1,4 @@
+{
+  "factory": "0x751A0bC0e3f75b38e01Cf25bFCE7fF36DE1C87DE",
+  "timestamp": 1626888045445
+}
