+++ pkg/deployments/tasks/20210721-liquidity-bootstrapping-pool/output/rinkeby.json
@@ -1,0 +1,4 @@
+{
+  "factory": "0xdcdbf71A870cc60C6F9B621E28a7D3Ffd6Dd4965",
+  "timestamp": 1626887161793
+}
