+++ pkg/deployments/tasks/20210721-liquidity-bootstrapping-pool/readme.md
@@ -1,0 +1,9 @@
+# 2021-07-21 - Liquidity Bootstrapping Pool
+
+Deployment of the `LiquidityBootstrappingPool`, for Liquidity Bootstrapping Pools of up to 4 tokens.
+
+## Useful Files
+
+- [Ethereum mainnet addresses](./output/mainnet.json)
+- [Polygon mainnet addresses](./output/polygon.json)
+- [`LiquidityBootstrappingPool` ABI](./abi/LiquidityBootstrappingPool.json)
